Move MessageBus interface/attributes to Messages
