+++ src/Framework/Infrastructure/DefaultMessageBus.php
@@ -5,6 +5,7 @@ namespace Smolblog\Framework\Infrastructure;
 use Crell\Tukio\Dispatcher;
 use Psr\EventDispatcher\EventDispatcherInterface;
 use Psr\EventDispatcher\ListenerProviderInterface;
+use Smolblog\Framework\Messages\MessageBus;
 use Smolblog\Framework\Messages\Query;
 
 /**
